% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gk_plot_dotplot}
\alias{gk_plot_dotplot}
\title{Plot marker means by classified cell type}
\usage{
gk_plot_dotplot(prepared, classification, thresholds)
}
\arguments{
\item{prepared}{A gk_prepare() result.}

\item{classification}{A gk_classify() result.}

\item{thresholds}{A gk_thresholds() result.}
}
\value{
A ggplot object.
}
\description{
Plot marker means by classified cell type
}
\seealso{
Other plots:
\code{\link[=gk_plot_density]{gk_plot_density()}},
\code{\link[=gk_plot_embedding]{gk_plot_embedding()}},
\code{\link[=gk_plot_map]{gk_plot_map()}},
\code{\link[=gk_plot_overview]{gk_plot_overview()}},
\code{\link[=gk_plot_states]{gk_plot_states()}},
\code{\link[=gk_plot_thresholds]{gk_plot_thresholds()}}
}
\concept{plots}
