% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gk_plot_map}
\alias{gk_plot_map}
\title{Plot cell centroids on an image map}
\usage{
gk_plot_map(x, image_id, colour_by = "cell_type", layers = NULL)
}
\arguments{
\item{x}{A gk_prepare(), gk_classify() or gk_replayed() result.}

\item{image_id}{Image to plot.}

\item{colour_by}{Cell column used for colour.}

\item{layers}{Reserved layer configuration.}
}
\value{
A ggplot object.
}
\description{
Plot cell centroids on an image map
}
\seealso{
Other plots:
\code{\link[=gk_plot_density]{gk_plot_density()}},
\code{\link[=gk_plot_dotplot]{gk_plot_dotplot()}},
\code{\link[=gk_plot_embedding]{gk_plot_embedding()}},
\code{\link[=gk_plot_overview]{gk_plot_overview()}},
\code{\link[=gk_plot_states]{gk_plot_states()}},
\code{\link[=gk_plot_thresholds]{gk_plot_thresholds()}}
}
\concept{plots}
