% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gk_plot_overview}
\alias{gk_plot_overview}
\title{Make a compact overview}
\usage{
gk_plot_overview(x, image_id)
}
\arguments{
\item{x}{A prepared, classified or replayed object.}

\item{image_id}{Image to plot.}
}
\value{
A patchwork object when patchwork is installed, otherwise the map.
}
\description{
Make a compact overview
}
\seealso{
Other plots:
\code{\link[=gk_plot_density]{gk_plot_density()}},
\code{\link[=gk_plot_dotplot]{gk_plot_dotplot()}},
\code{\link[=gk_plot_embedding]{gk_plot_embedding()}},
\code{\link[=gk_plot_map]{gk_plot_map()}},
\code{\link[=gk_plot_states]{gk_plot_states()}},
\code{\link[=gk_plot_thresholds]{gk_plot_thresholds()}}
}
\concept{plots}
