% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gk_plot_states}
\alias{gk_plot_states}
\title{Plot state calls for one image}
\usage{
gk_plot_states(state_calls, image_id)
}
\arguments{
\item{state_calls}{A gk_state_calls() result.}

\item{image_id}{Image to plot.}
}
\value{
A ggplot object.
}
\description{
Plot state calls for one image
}
\seealso{
Other plots:
\code{\link[=gk_plot_density]{gk_plot_density()}},
\code{\link[=gk_plot_dotplot]{gk_plot_dotplot()}},
\code{\link[=gk_plot_embedding]{gk_plot_embedding()}},
\code{\link[=gk_plot_map]{gk_plot_map()}},
\code{\link[=gk_plot_overview]{gk_plot_overview()}},
\code{\link[=gk_plot_thresholds]{gk_plot_thresholds()}}
}
\concept{plots}
