% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{gk_read_export}
\alias{gk_read_export}
\title{Read a verified export}
\usage{
gk_read_export(dir, verify = TRUE)
}
\arguments{
\item{dir}{Export directory.}

\item{verify}{Verify before reading?}
}
\value{
A \code{gk_export} list containing snapshot, configuration and tables.
}
\description{
Read a verified export
}
\seealso{
Other export:
\code{\link[=gk_export]{gk_export()}},
\code{\link[=gk_lock]{gk_lock()}},
\code{\link[=gk_verify_export]{gk_verify_export()}}
}
\concept{export}
